requirements.txt
asgiref==3.7.2
boto3==1.34.34
botocore==1.34.34
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.60.1
django-environ==0.11.2
django-storages==1.14.2
facebook-sdk==3.1.0
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tweepy==4.14.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
